% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_productivity_index}
\alias{get_productivity_index}
\title{Function that calculates a productivity index (weighted mean) for a given time period}
\usage{
get_productivity_index(key, dateFrom, dateTo)
}
\arguments{
\item{key}{string provided by RescueTime}

\item{dateFrom}{string in iso date format "YYYY-mm-DDTHH:MM:ssZ"}

\item{dateTo}{string in iso date format "YYYY-mm-DDTHH:MM:ssZ"}
}
\value{
data.frame with a productivity for each date
}
\description{
Function that calculates a productivity index (weighted mean) for a given time period
}
\examples{
productivity_index <- get_productivity_index("XYZ","2021-04-11T00:00:00Z","2021-04-12T23:59:59Z")
}
