% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_rescue_time_data}
\alias{get_rescue_time_data}
\title{Get RescueTime Data based on an API Key.
The Data is returned unfiltered}
\usage{
get_rescue_time_data(
  key,
  dateFrom,
  dateTo,
  scope = "Activity",
  transform = FALSE
)
}
\arguments{
\item{key}{string provided by RescueTime}

\item{dateFrom}{string in iso date format "YYYY-mm-DDTHH:MM:ssZ"}

\item{dateTo}{string in iso date format "YYYY-mm-DDTHH:MM:ssZ"}

\item{scope}{string either "Overview", "Category", "Activity", "Productivity", "Document"}

\item{transform}{boolean transform data to different structure}
}
\value{
table with complete data
}
\description{
Get RescueTime Data based on an API Key.
The Data is returned unfiltered
}
\examples{
data <- get_rescue_time_data("XYZ","2021-04-11T00:00:00Z","2021-04-12T23:59:59Z","Category")
}
