% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_rescue_time_data_anonymized}
\alias{get_rescue_time_data_anonymized}
\title{Get RescueTime Data based on an API Key
The Data is return anonymized according to the category definitions on package level}
\usage{
get_rescue_time_data_anonymized(
  key,
  dateFrom,
  dateTo,
  scope = "Category",
  transform = FALSE
)
}
\arguments{
\item{key}{string provided by RescueTime}

\item{dateFrom}{string in iso date format "YYYY-mm-DDTHH:MM:ssZ"}

\item{dateTo}{string in iso date format "YYYY-mm-DDTHH:MM:ssZ"}

\item{scope}{string either "Category" or "SubCategory"}

\item{transform}{boolean transform data to different structure}
}
\value{
table with an entry for each category for each day
}
\description{
Get RescueTime Data based on an API Key
The Data is return anonymized according to the category definitions on package level
}
\examples{
data <- get_rescue_time_data_anonymized("XYZ","2021-04-11T00:00:00Z","2021-04-12T23:59:59Z","Category")
}
