% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{transform_data}
\alias{transform_data}
\title{Transform data frame based on row and column definitions}
\usage{
transform_data(data, row_column, col_column, value_columns = c("Time"))
}
\arguments{
\item{data}{data.frame with rows and columns}

\item{row_column}{string the column that contains the values for the transformed row}

\item{col_column}{string the column that contains the values for the transformed columns}

\item{ignored_columns}{array list of columns to ignore as value columns}
}
\value{
a data frame that has a row for each entry of the values in the row_column
and columns for each value of the col_column
}
\description{
Transform data frame based on row and column definitions
}
\examples{
new_data <- transform_data(data, "Date", "Category")
}
